// server/lib/provider/responses/openai-responses-api.ts
export type TeamAIResponsesTool =
  | {
      type: 'function'
      name: string
      description: string | undefined
      parameters: JSONSchema7
      strict: boolean | undefined
    }
  | {
      type: 'web_search'
      max_keyword?: number
      limit?: number
      sources: ('toutiao' | 'douyin' | 'moji')[] | undefined
      user_location?: {
        type: 'approximate'
        city?: string
        country?: string
        region?: string
      }
    }